Bugfix: Use reflection to build ParametersByCategory dictionary instead of humongous manual list.
Caching the result for performance.

This fixes strange compiler errors ("csc.dll exited with code 138"). Possibly due to memory restrictions when compiling the humongous ImmutableList.
Reason: Each Add() on the ImmutableList created a new copy.
Quote from the docs:
"You can then call methods, such as Add and AddRange, to populate the collection. Note that these methods return a new object. When you add or remove items from an immutable list, a copy of the original list is made with the items added or removed, and the original list is unchanged."